Updated files and added product images

+++ Cart/models.py
@@ -74,7 +74,7 @@ class Cart(models.Model):
 class CartItem(models.Model):
     cart = models.ForeignKey(Cart, on_delete=models.CASCADE)
     product = models.ForeignKey(Product, on_delete=models.CASCADE)
-    quantity = models.PositiveBigIntegerField(default=1)
+    quantity = models.PositiveIntegerField(default=1)
     product_size_color = models.ForeignKey(ProductSizeColor, on_delete=models.CASCADE, null=True)
